// src/app/models/Employee.ts
export class Employee{
  empId: string;
  firstname: string;
  lastName: string;
  email: string;
  address: string;
  imageUrl: string;
  gender: string;
  phoneNumber: 0;
  salary: 0;
    designationName: string;
    designations: {
      designationName: string;
      roleName: string;
      departmentName: string;
    }
}